[wxOSX] Keep native font when changing wxStaticBox label colour
On mac wxStaticBox::SetFont() preserves both the native font name and the size of the label text.
When setting the foreground colour however, the font was set to the one inherited from the parent window.
Override SetForegroundColour in wxStaticBox on mac and tweak the code to set the colour only by passing a dummy font object